# internal/domain/dto/booking_dto.go
package dto

type BookingRequestDto struct {
	GuestId        string `json:"mainGuest"`
	NumberOfGuests int    `json:"numberOfGuests"`
	CheckInDate    string `json:"checkInDate"`
	CheckOutDate   string `json:"checkOutDate"`
}

type ConfirmationDto struct {
	Message   string              `json:"message"`
	BookingId string              `json:"bookingId"`
	Info      ConfirmationInfoDto `json:"info"`
}

type ConfirmationInfoDto struct {
	CottageName    string `json:"cottageName"`
	NumberOfGuests int    `json:"numberOfGuests"`
	CheckInDate    string `json:"checkInDate"`
	CheckOutDate   string `json:"checkOutDate"`
}